mem_dialog_1/requirements.txt
streamlit==1.32.2
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.3
scikit-learn==1.4.2
Pillow==10.2.0
